+++ resolver_test.go
@@ -39,43 +39,27 @@ var _ = Describe("Resolver", func() {
 			output := Verify(names, m)
 			var verified, notVerified int
 			for _, o := range output {
-				if o.Verified {
-					verified++
-				} else {
+				if o.MatchType == "NoMatch" {
 					notVerified++
+				} else {
+					verified++
 				}
 			}
 			Expect(verified).To(Equal(6))
 			Expect(notVerified).To(Equal(2))
 		})
+
+		It("has expected fields", func() {
+			m := util.NewModel()
+			name := "Homo sapiens"
+			nameOutputs := Verify([]string{name}, m)
+			result := nameOutputs[name]
+			Expect(result.DatabasesNum).To(BeNumerically(">", 2))
+			Expect(result.MatchType).To(Equal("ExactCanonicalMatch"))
+			Expect(result.DataSourceID).To(BeNumerically(">", 0))
+			Expect(result.MatchedName).To(Equal("Homo sapiens Linnaeus, 1758"))
+			Expect(result.ClassificationPath).To(Equal("Animalia|Chordata|Mammalia|Primates|Hominoidea|Hominidae|Homo|Homo sapiens"))
+			Expect(result.CurrentName).To(Equal("Homo sapiens Linnaeus, 1758"))
+		})
 	})
 })
-
-// 	Describe("Run", func() {
-
-// 		It("has all fields", func() {
-// 			m := util.NewModel()
-// 			name := "Homo sapiens"
-// 			nameOutputs := <-Run([]string{name}, m)
-// 			result := nameOutputs[name]
-// 			Expect(result.Resolved).To(BeTrue())
-// 			Expect(result.Total).To(Equal(8))
-// 			// Expect(result.MatchType).To(Equal("Match")) // pending
-// 			Expect(result.DataSourceId).To(Equal(1))
-// 			Expect(result.Name).To(Equal("Homo sapiens Linnaeus, 1758"))
-// 			Expect(result.ClassificationPath).To(Equal("Animalia|Chordata|Mammalia|Primates|Hominoidea|Hominidae|Homo|Homo sapiens"))
-// 			Expect(result.AcceptedName).To(Equal("Homo sapiens Linnaeus, 1758"))
-// 		})
-
-// 		It("handles simple and advanced resolution", func() {
-// 			m := util.NewModel()
-// 			name := "Homo sapiens"
-// 			nameOutputs := <-Run([]string{name}, m)
-// 			// Expect(nameOutputs[name].MatchType).To(Equal("ExactCanonicalMatch")) // pending
-
-// 			m.Resolver.AdvancedResolution = true
-// 			nameOutputs = <-Run([]string{name}, m)
-// 			Expect(nameOutputs[name].MatchType).To(Equal("ExactCanonicalMatch"))
-// 		})
-// 	})
-// })